requirements/base.txt
fastapi
pydantic-settings
pydantic
uvicorn
tortoise-orm[asyncpg]
aerich
fastapi-limiter
aioredis
aiohttp
aio-celery[redis]
openai
loguru